Refactoring of HistoryViewModel
