+++ App/App/ViewModels/HistoryViewModel.cs
@@ -8,6 +8,7 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using Xamarin.Forms;
 
@@ -24,6 +25,14 @@ namespace App.ViewModels
 		private readonly StatisticsManager _stats = DependencyService.Get<StatisticsManager>();
 
 		public readonly int[] MonthAndDay = { DateTime.Now.Month, DateTime.Now.Day };
+
+		public ObservableCollection<MovementItemViewModel> Movements = new ObservableCollection<MovementItemViewModel>();
+
+		public bool ShowEmptyLabel => Movements.Count == 0 && !FirstLoad;
+
+		public bool FirstLoad { get; set; } = true;
+
+		public string SortIconSource => SortOrder == SortOrder.Ascending ? DESCENDING_ICON_SOURCE : ASCENDING_ICON_SOURCE;
 
 		private int _year;
 		public int Year
@@ -49,14 +58,6 @@ namespace App.ViewModels
 			get => _isRefreshing;
 			set => SetProperty(ref _isRefreshing, value);
 		}
-
-		public ObservableCollection<MovementItemViewModel> Movements = new ObservableCollection<MovementItemViewModel>();
-
-		public bool ShowEmptyLabel => Movements.Count == 0 && !FirstLoad;
-
-		public bool FirstLoad { get; set; } = true;
-
-		public string SortIconSource => SortOrder == SortOrder.Ascending ? DESCENDING_ICON_SOURCE : ASCENDING_ICON_SOURCE;
 
 		private SortOrder _sortOrder;
 		public SortOrder SortOrder
@@ -103,7 +104,7 @@ namespace App.ViewModels
 		{
 			await Filter(
 				new DateTime(_year, 1, 1),
-				x => x.CreationDate.Year == _year,
+				m => m.CreationDate.Year == _year,
 				SearchDepth.Year).ConfigureAwait(false);
 		}
 
@@ -111,7 +112,7 @@ namespace App.ViewModels
 		{
 			await Filter(
 				new DateTime(_year, MonthAndDay[0], 1),
-				x => x.CreationDate.Year == _year && x.CreationDate.Month == MonthAndDay[0],
+				m => m.CreationDate.Year == _year && m.CreationDate.Month == MonthAndDay[0],
 				SearchDepth.Month).ConfigureAwait(false);
 		}
 
@@ -120,7 +121,7 @@ namespace App.ViewModels
 			var date = new DateTime(_year, MonthAndDay[0], MonthAndDay[1]);
 			await Filter(
 				date,
-				x => x.CreationDate.Date == date.Date,
+				m => m.CreationDate.Date == date.Date,
 				SearchDepth.Day).ConfigureAwait(false);
 		}
 
@@ -136,30 +137,14 @@ namespace App.ViewModels
 				if (data.Length == 0)
 					return;
 
-				var index = GetStartingIndex(data, date, depth);
+				var index = FilterHelpers.GetStartingIndex(data, date, depth);
 				if (index == -1)
 					return;
 
 				if (SortOrder == SortOrder.Ascending)
-				{
-					for (; index < data.Length; index++)
-					{
-						if (filterCondition(data[index]))
-							Movements.Add(new MovementItemViewModel(data[index]));
-						else
-							return;
-					}
-				}
+					InsertAscending(index, data, filterCondition);
 				else
-				{
-					for (; index < data.Length; index++)
-					{
-						if (filterCondition(data[index]))
-							Movements.Insert(0, new MovementItemViewModel(data[index]));
-						else
-							return;
-					}
-				}
+					InsertDescending(index, data, filterCondition);
 			}
 			catch (Exception ex)
 			{
@@ -171,67 +156,26 @@ namespace App.ViewModels
 			}
 		}
 
-		private int GetStartingIndex(Movement[] items, DateTime toFind, SearchDepth depth)
+		private void InsertAscending(int index, Movement[] movementsToAdd, Func<Movement, bool> condition)
 		{
-			var index = -1;
-
-			switch (depth)
+			for (; index < movementsToAdd.Length; index++)
 			{
-				case SearchDepth.Year:
-					index = BinarySearchDate(items, toFind.Year, 0, items.Length);
-					break;
-				case SearchDepth.Month:
-					index = BinarySearchDate(items, toFind.Year, toFind.Month, 0, items.Length);
-					break;
-				case SearchDepth.Day:
-					index = BinarySearchDate(items, toFind, 0, items.Length);
-					break;
+				if (condition(movementsToAdd[index]))
+					Movements.Add(new MovementItemViewModel(movementsToAdd[index]));
+				else
+					return;
 			}
-
-			while (index >= 0 && items[index].CreationDate.Date >= toFind.Date)
-				index--;
-
-			return index > 0
-				? ++index
-				: items[0].CreationDate.Date >= toFind.Date
-					? 0
-					: -1;
 		}
 
-		private int BinarySearchDate(Movement[] items, int year, int low, int high)
+		private void InsertDescending(int index, Movement[] movementsToAdd, Func<Movement, bool> condition)
 		{
-			if (low >= high)
-				return -1;
-			var mid = (low + high - 1) / 2;
-			if (items[mid].CreationDate.Year == year)
-				return mid;
-			if (year < items[mid].CreationDate.Year)
-				return BinarySearchDate(items, year, low, mid - 1);
-			return BinarySearchDate(items, year, mid + 1, high);
-		}
-
-		private int BinarySearchDate(Movement[] items, int year, int month, int low, int high)
-		{
-			if (low >= high)
-				return -1;
-			var mid = (low + high - 1) / 2;
-			if (items[mid].CreationDate.Year == year && items[mid].CreationDate.Month == month)
-				return mid;
-			if (year < items[mid].CreationDate.Year || month < items[mid].CreationDate.Month)
-				return BinarySearchDate(items, year, month, low, mid - 1);
-			return BinarySearchDate(items, year, month, mid + 1, high);
-		}
-
-		private int BinarySearchDate(Movement[] items, DateTime date, int low, int high)
-		{
-			if (low >= high)
-				return -1;
-			var mid = (low + high - 1) / 2;
-			if (items[mid].CreationDate.Date == date.Date)
-				return mid;
-			if (date.Date < items[mid].CreationDate.Date)
-				return BinarySearchDate(items, date, low, mid - 1);
-			return BinarySearchDate(items, date, mid + 1, high);
+			for (; index < movementsToAdd.Length; index++)
+			{
+				if (condition(movementsToAdd[index]))
+					Movements.Insert(0, new MovementItemViewModel(movementsToAdd[index]));
+				else
+					return;
+			}
 		}
 	}
 }
